[12.0][ADD] - Improve multi-company management for brand model
In a multi-company management system, there is a strong link between the company
and its brands. A company can't make sells under a brand that it don't own.
For this reason we add a mixin to be inherited from branded objects that assure the coherence of this link.

We also add an multi-company access rule for brand model.

A new field added to the company to define if the brand field is required or optional by company.