analytic_brand: Run tests in post-install:
There is a common issue in v18 when creating a new res.partner in tests
(which is the case here as we create a res.brand).
There is a required field autopost_bills in account addon.
As this addon doesn't depend on account, if we install this addon
on a DB where account is also installed, tests will fail because
the autopost_bills column is created into postgres, with the not null constraint,
but the default value won't be set by Odoo.

Running these tests in post-install ensure Odoo will set the default value
as specified in the account addon.